requirements.txt
discord.py==1.7.3
aiohttp
websockets
chardet
pycparser